Check message types
