+++ src/utils/mtproto/constants.js
@@ -34,3 +34,4 @@ export const GET_NEAREST_DC = 0x1fb33026;
 export const HTTP_WAIT = 0x9299359f;
 
 export const MESSAGE_CONTAINER = 0x73f1f8dc;
+export const NEW_SESSION_CREATED = 0x9ec20908;
